backend/pyproject.toml
[tool.poetry]
name = "agentic-rag"
version = "0.1.0"
description = "A sophisticated multi-agent RAG system using LangGraph and Google Gemini"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "agents"}, {include = "api"}, {include = "config"}, {include = "graph"}, {include = "models"}, {include = "tools"}, {include = "utils"}]

[tool.poetry.dependencies]
python = "^3.10"
# Core dependencies
python-dotenv = "^1.0.1"
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"

# LangChain Core (langchain-core >=1.0.0 required for langchain-google-genai 3.x)
langchain = ">=1.0.0"
langchain-core = ">=1.0.0"
langchain-community = ">=0.3.0"
langchain-text-splitters = ">=0.3.0"

# LangGraph for Multi-Agent Orchestration
# Note: langchain 1.x requires langgraph >=1.0.0
langgraph = ">=1.0.0"

# Google Gemini Integration
langchain-google-genai = "^3.0.0"

# OpenAI Integration (for embeddings)
langchain-openai = "^0.3.0"

# Vector Store (FAISS for local development)
faiss-cpu = "^1.8.0"

# Web Search Tool
tavily-python = "^0.5.0"

# Document Processing
pypdf = "^5.0.0"
python-docx = "^1.1.0"

# Text Processing
tiktoken = "^0.8.0"

# Utilities
aiofiles = "^24.1.0"
httpx = "^0.27.0"

# Streaming & Persistence
sse-starlette = "^1.8.0"
langgraph-checkpoint-sqlite = "^2.0.0"
langgraph-checkpoint-postgres = "^2.0.0"
aiosqlite = "^0.20.0"
psycopg = {extras = ["binary", "pool"], version = "^3.2.0"}

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
